requirements.txt
datasets==4.4.2
huggingface_hub==0.36.0
PyYAML==6.0.2
torch==2.9.0
tqdm==4.67.1
transformers==4.57.3
vllm==0.13.0